Added some general commenting to StatHolderScript.cs
Added some general commenting to StatHolderScript.cs. This file is was not written with the best OOP practices outside of unity. However, It fills its role in dynamically creating and managing all dot and input controls for the character sheet. At this point, I was worried about time. So I pushed through linier logic to accomplish the underlaying goal of character sheet generation/loading/manipulation.